move mosaic to its own package
