+++ setup.cfg
@@ -26,6 +26,16 @@ replace = version="{new_version}"
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
+
+[bumpversion:titiler/mosaic:setup.py]
+search = version="{current_version}"
+replace = version="{new_version}"
+
+[bumpversion:file:titiler/mosaic/titiler/mosaic/version.py]
+search = __version__ = "{current_version}"
+replace = __version__ = "{new_version}"
+
+
 # Linter configs
 [isort]
 profile=black
