+++ titiler/application/titiler/application/routers/mosaic.py
@@ -1,7 +1,7 @@
 """TiTiler MosaicJSON demo endpoints."""
 
 from titiler.application.custom import ColorMapParams
-from titiler.core.factory import MosaicTilerFactory
+from titiler.mosaic.factory import MosaicTilerFactory
 
 mosaic = MosaicTilerFactory(
     colormap_dependency=ColorMapParams, router_prefix="mosaicjson"
