+++ titiler/core/titiler/core/errors.py
@@ -3,15 +3,8 @@
 import logging
 from typing import Callable, Dict, Type
 
-from cogeo_mosaic.errors import (
-    MosaicAuthError,
-    MosaicError,
-    MosaicNotFoundError,
-    NoAssetFoundError,
-)
 from rasterio.errors import RasterioError, RasterioIOError
 from rio_tiler.errors import (
-    EmptyMosaicError,
     InvalidAssetName,
     InvalidBandName,
     InvalidColorFormat,
@@ -44,19 +37,14 @@ class BadRequestError(TilerError):
 
 DEFAULT_STATUS_CODES = {
     BadRequestError: status.HTTP_400_BAD_REQUEST,
-    MosaicAuthError: status.HTTP_401_UNAUTHORIZED,
     TileOutsideBounds: status.HTTP_404_NOT_FOUND,
     TileNotFoundError: status.HTTP_404_NOT_FOUND,
-    EmptyMosaicError: status.HTTP_404_NOT_FOUND,
-    MosaicNotFoundError: status.HTTP_404_NOT_FOUND,
-    NoAssetFoundError: status.HTTP_404_NOT_FOUND,
     RasterioIOError: status.HTTP_404_NOT_FOUND,
     MissingBands: status.HTTP_400_BAD_REQUEST,
     MissingAssets: status.HTTP_400_BAD_REQUEST,
     InvalidColorFormat: status.HTTP_400_BAD_REQUEST,
     InvalidAssetName: status.HTTP_404_NOT_FOUND,
     InvalidBandName: status.HTTP_404_NOT_FOUND,
-    MosaicError: status.HTTP_424_FAILED_DEPENDENCY,
     RasterioError: status.HTTP_500_INTERNAL_SERVER_ERROR,
     RioTilerError: status.HTTP_500_INTERNAL_SERVER_ERROR,
     Exception: status.HTTP_500_INTERNAL_SERVER_ERROR,
