+++ titiler/mosaic/setup.py
@@ -1,4 +1,4 @@
-"""Setup titiler-core."""
+"""Setup titiler-mosaic."""
 
 from setuptools import find_namespace_packages, setup
 
@@ -6,20 +6,8 @@ with open("README.md") as f:
     long_description = f.read()
 
 inst_reqs = [
-    "fastapi==0.63.0",
-    "geojson-pydantic",
-    "jinja2>=2.11.2,<3.0.0",
-    "morecantile>=2.1,<2.2",
-    "numpy",
-    "pydantic",
-    "rasterio",
-    "rio-cogeo>=2.1,<2.2",
-    "rio-tiler>=2.0.5,<2.1",
-    # Additional requirements for python 3.6
-    "async_exit_stack>=1.0.1,<2.0.0;python_version<'3.7'",
-    "async_generator>=1.10,<2.0.0;python_version<'3.7'",
-    "dataclasses;python_version<'3.7'",
-    "importlib_resources;python_version<'3.7'",
+    "titiler-core",
+    "cogeo-mosaic>=3.0.0rc2,<3.1",
 ]
 extra_reqs = {
     "test": ["pytest", "pytest-cov", "pytest-asyncio", "requests"],
@@ -27,9 +15,9 @@ extra_reqs = {
 
 
 setup(
-    name="titiler-core",
+    name="titiler-mosaic",
     version="0.2.0",
-    description=u"A modern dynamic tile server built on top of FastAPI and Rasterio/GDAL.",
+    description=u"MosaicJSON plugin for TiTiler.",
     long_description=long_description,
     long_description_content_type="text/markdown",
     python_requires=">=3.6",
@@ -40,14 +28,14 @@ setup(
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
     ],
-    keywords="COG STAC FastAPI",
+    keywords="MosaicJSON",
     author=u"Vincent Sarago",
     author_email="vincent@developmentseed.org",
     url="https://github.com/developmentseed/titiler",
     license="MIT",
     packages=find_namespace_packages(exclude=["tests*"]),
-    package_data={"titiler": ["core/templates/*.xml"]},
     include_package_data=True,
     zip_safe=False,
     install_requires=inst_reqs,
