+++ titiler/mosaic/tests/conftest.py
@@ -1,0 +1,27 @@
+"""``pytest`` configuration."""
+
+import os
+from typing import Any, Dict
+
+import pytest
+from rasterio.io import MemoryFile
+
+DATA_DIR = os.path.join(os.path.dirname(__file__), "fixtures")
+
+
+@pytest.fixture
+def set_env(monkeypatch):
+    """Set Env variables."""
+    monkeypatch.setenv("AWS_ACCESS_KEY_ID", "jqt")
+    monkeypatch.setenv("AWS_SECRET_ACCESS_KEY", "rde")
+    monkeypatch.setenv("AWS_DEFAULT_REGION", "us-west-2")
+    monkeypatch.setenv("AWS_REGION", "us-west-2")
+    monkeypatch.delenv("AWS_PROFILE", raising=False)
+    monkeypatch.setenv("AWS_CONFIG_FILE", "/tmp/noconfigheere")
+
+
+def parse_img(content: bytes) -> Dict[Any, Any]:
+    """Read tile image and return metadata."""
+    with MemoryFile(content) as mem:
+        with mem.open() as dst:
+            return dst.profile
