+++ titiler/mosaic/tests/fixtures/mosaic.json
@@ -1,0 +1,1 @@
+{"mosaicjson": "0.0.2", "quadkey_zoom": 7, "minzoom": 7, "maxzoom": 9, "bounds": [-75.98703377403767, 44.93504283303786, -71.337604724099, 47.096855991923235], "center": [-74.53125, 45.99569351896393, 7], "tiles": {"0302300": ["cog1.tif"], "0302301": ["cog1.tif", "cog2.tif"], "0302310": ["cog1.tif", "cog2.tif"], "0302302": ["cog1.tif"], "0302303": ["cog1.tif", "cog2.tif"], "0302312": ["cog1.tif", "cog2.tif"], "0302320": ["cog1.tif"], "0302321": ["cog1.tif", "cog2.tif"], "0302330": ["cog1.tif", "cog2.tif"]}}
