+++ titiler/mosaic/titiler/mosaic/__init__.py
@@ -1,0 +1,5 @@
+"""titiler.mosaic"""
+
+from . import errors, factory, version  # noqa
+from .factory import MosaicTilerFactory  # noqa
+from .version import __version__  # noqa
