+++ titiler/mosaic/titiler/mosaic/version.py
@@ -1,0 +1,3 @@
+"""titiler.mosaic version."""
+
+__version__ = "0.2.0"
