+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py36,py37,py38
+envlist = py36,py37,py38,py39
 
 [testenv]
 deps =
@@ -10,31 +10,9 @@ deps =
 skip_install = true
 commands_pre =
     python -m pip install titiler/core
+    python -m pip install titiler/mosaic
     python -m pip install titiler/application
 commands =
     python -m pytest titiler/core --cov titiler.core --cov-report xml --cov-report term-missing --ignore=venv
+    python -m pytest titiler/mosaic --cov titiler.mosaic --cov-report xml --cov-report term-missing --ignore=venv
     python -m pytest titiler/application --cov titiler.application --cov-report xml --cov-report term-missing --ignore=venv
-
-# Release tooling
-[testenv:build]
-basepython = python3
-skip_install = true
-deps =
-    wheel
-    setuptools
-commands =
-    python setup.py sdist
-
-[testenv:release]
-setenv =
-    TWINE_USERNAME = {env:TWINE_USERNAME}
-    TWINE_PASSWORD = {env:TWINE_PASSWORD}
-basepython = python3
-skip_install = true
-deps =
-    {[testenv:build]deps}
-    twine >= 1.5.0
-commands =
-    {[testenv:build]commands}
-    twine upload --skip-existing dist/*
-
